requirements.txt
anthropic
fastapi
uvicorn[standard]
apscheduler
chromadb
sentence-transformers
python-dotenv
pydantic